Pull out some helper functions
